roll_storage_backend/requirements.txt
asgiref==3.7.2
colorama==0.4.6
Django==3.2.3
django-pytest==0.2.0
djangorestframework==3.14.0
exceptiongroup==1.1.3
flake8==6.1.0
gunicorn==20.1.0
iniconfig==2.0.0
mccabe==0.7.0
packaging==23.2
pluggy==1.3.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pycodestyle==2.11.0
pyflakes==3.1.0
pytest==7.4.2
pytest-django==4.5.2
pytz==2023.3.post1
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.8.0
